shells/backends/azure_tf_backend/src/requirements.txt
cloudshell-shell-core>=5.0.3,<6.0.0
azure-identity == 1.0.0
azure-storage-blob == 12.8.1
